feat: sets User Agent on API requests
The user agent is set on all API requests, and includes the mailto so that
requests are routed to the Polite API pool.

Requests for citation information use a mailto prepended on the request
url, as these are made by Javascript, where it's not possible to override
the User-Agent header.